server/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pandas
python-multipart
dvc
evidently
langdetect
pdfkit
pillow
jinja2
